uvsensor/requirements.txt
asyncio
aiofiles
aiohttp
argparse
slack_sdk
slackeventsapi
pyyaml
numpy
pandas
matplotlib
adafruit-circuitpython-ads1x15
